Encode senone_active as deltas and use fewer bits (in fact we can use 8 bits and just eat the occasional extra senone...)
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx@8744 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libpocketsphinx/ms_mgau.h
@@ -144,7 +144,7 @@ void ms_mgau_free(ms_mgau_model_t *g /**< In: A set of models to free */
 
 int32 ms_cont_mgau_frame_eval(ms_mgau_model_t * msg,
                               int16 *senscr,
-                              int32 *senone_active,
+                              int16 *senone_active,
                               int32 n_senone_active,
                               mfcc_t ** feat,
                               int32 frame,
